+++ sender-receiver-linux/Logging/LogEntries/PendulumLogEntry.cpp
@@ -33,7 +33,7 @@ PendulumLogEntry::PendulumLogEntry(time_point<system_clock> timePoint, unsigned 
 
 // JSON converter class (used implicitely by nlohmann/json.hpp):
 void to_json(nlohmann::json &jsonObject, const PendulumLogEntry &entry) {
-    long timePointAsUnixMillis = std::chrono::duration_cast<std::chrono::milliseconds>(
+    unsigned long long timePointAsUnixMillis = std::chrono::duration_cast<std::chrono::milliseconds>(
             entry.getTimePoint().time_since_epoch()).count();
 
     jsonObject = {
